+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
-## [UNRELEASED]
+## [2021.11.1] 2021.11.1
 ### Changed
-- `clonedb_whitelist.sh` will not try to write out mapfiles, mapcachefile and QGIS files.
+- `clonedb_whitelist.sh` will not try to write out mapfiles, mapcachefile and QGIS files. Use gc2-cli for this.
+- API tests now uses $BUILD_ID env var if set for usernames, so they can be predicted in Vidi tests.
 
 ## [2021.11.0] - 2021-3-11
 ### Fixed